requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
albumentations>=1.4.4
opencv-python>=4.10
pandas>=2.2
numpy>=1.26
matplotlib>=3.9
scikit-learn>=1.5
tqdm>=4.66
pyyaml>=6.0
gradio>=4.44
huggingface-hub>=0.25
transformers>=4.44
accelerate>=0.34
faiss-cpu>=1.7.4
sentence-transformers>=3.0.1
onnxruntime>=1.19
einops>=0.8
nbformat>=5.10
